libgeda: Create interned strings for attribute names
There are a lot of places in geda-gaf where attribute names are
compared, usually using `strcmp()`.  In reality, the set of attribute
names that are likely to be encountered in a design is very small, so
in many cases it would be a valid optimisation to use an interned
string to represent the name, since the pointers can then be compared
directly.

Text objects only have their text set via `o_text_set_string()`.  This
calls `o_text_recreate()`, which in turn calls `update_disp_string()`,
which parses the name and value from the string.  By caching an
interned string in the text info structure, almost all calls to
`o_attrib_string_get_name_value()` can be removed -- saving a lot of
CPU cycles spent on string processing.