gui: Start Multiattrib dialog on Edit -> Edit... for primitives
Several primitives were not processed so far in the code that
selects an edit action depending on the object type.  As any
schematic primitive may have attributes attached to it, it is
better to open the Multiattrib dialog instead of simply ignoring
the user input.  The code has been fixed to open the dialog in
such cases.